IMPALA-11232: Do not add some jars to HADOOP_CLASSPATH when starting HMS
This patch changes the line that added to HADOOP_CLASSPATH all the jar
files in the folder ${RANGER_HOME}/ews/webapp/WEB-INF/lib to a line that
only includes those jar files with names starting with "ranger-" since
almost all other jar files do not seem to be necessary to run the E2E
test of test_hive_with_ranger_setup.

This way we also avoid adding too many paths to HADOOP_CLASSPATH, which
in turn could result in Hadoop not being able to return its version to
the script that starts HMS due to the error of "Argument list too long".

Testing:
 - Verified after this patch, test_hive_with_ranger_setup still
   succeeds.
 - Verified in a local development environment that the length of
   Hadoop's environment variable 'CLASSPATH' logged in
   hive-metastore.out decreases from 100,876 characters to 62,634
   characters when executing run-hive-server.sh with the flag
   '-with_ranger' if $HADOOP_SHELL_SCRIPT_DEBUG is "true" and
   $IMPALA_HOME is "/home/fangyurao/Impala_for_FE".

Change-Id: Ifd66fd99a346835b9f81f95b5f046273fcce2590
Reviewed-on: http://gerrit.cloudera.org:8080/18398
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>