IMPALA-11236: Upgrade ehcache sizeof library to 0.4.0
It was found that the ehcache sizeof (version 0.3.0) library used by
CatalogdMetaProvider can fail due to an InaccessibleObjectException
during reflection and cause the impalad to stuck during startup
indefinitely. This could trigger IMPALA-7168 which affects sync_ddl
queries running on unaffected impalads.

This patch upgrades the version of the ehcache sizeof library so that
the exception could be handled within the ehcache sizeof library after
the upgrade.

Testing:
 - Verified that the patch passes Impala's core tests.

Change-Id: I853264684aa2ba9c3743f9e53bd52fa2752a45a1
Reviewed-on: http://gerrit.cloudera.org:8080/18400
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Quanlong Huang <huangquanlong@gmail.com>
Tested-by: Quanlong Huang <huangquanlong@gmail.com>