IMPALA-11291: deflake minidump-test by not starting JVM
minidump-test is unified into unified-be-test which will always init the
JVM and start the JVM pause monitor. It will print the following log to
stdout:
22/05/16 11:35:46 INFO util.JvmPauseMonitor: Starting JVM pause monitor

It's printed concurrently with the minidump output, which makes the test
flaky since it's verifying the output lines.

This patch moves minidump-test out of unified-be-test and explicitly
invokes InitCommonRuntime with init_jvm=false. So it won't be disrupted
by the above log.

Tests
 - The issue occurs when I run minidump-test several times repeatedly.
   After the fix, I can run minidump-test more than 2000 times without
   errors.

Change-Id: I89f81d408a2e905d5dfdd7f87177ebe2079d4d27
Reviewed-on: http://gerrit.cloudera.org:8080/18529
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>