IMPALA-11295: Deflake TestParquet.test_multiple_blocks_mt_dop
TestParquet.test_multiple_blocks_mt_dop runs a query on 6 scan ranges
using mt_dop=2. It then verifies the sum of ranges read on a backend is
2 (6/3). The test assumes that counters of the 2 instances on the same
host are printed consecutively. However, this is not always true. They
could be interleaving.

This patch makes the test more robust by grouping the counters based on
the host.

Test
 - I can't reproduce the issue locally. But I'm able to run the new test
   100 times without any error.

Change-Id: I16c576c41a212f83dda82a83931ab336a78a41e4
Reviewed-on: http://gerrit.cloudera.org:8080/18533
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-on: http://gerrit.cloudera.org:8080/18892
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>
Tested-by: Quanlong Huang <huangquanlong@gmail.com>