[tools] Fix building shell_pypi_package
Fixes an error building shell_pypi_package with junit_xml_wrapper.sh.
Previously `make shell_pypi_package` would error with

  /home/michael/Impala/bin/junitxml_command_wrapper.sh: line 46:
  DIST_DIR=/home/michael/Impala/shell/dist CLEAN_DIST=true
  /home/michael/Impala/shell/packaging/make_python_package.sh:
  No such file or directory

Updates the default `DIST_DIR` to point to the documented path that
CMake sets it to, and removes setting `DIST_DIR` as it matches the
default value. Also removes `CLEAN_DIST` as that value is ignored.

Change-Id: I60ffac3edf1a6027afa4ca46ab6dadfc6bfc660a
Reviewed-on: http://gerrit.cloudera.org:8080/18578
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>