IMPALA-11342: Fix class loading in Hive UDFs' constructors
Loading new classes from the same jar in the constructor of UDFs
did not work in the catalog because the URLClassLoader was closed
too early. Extended the lifecycle of the class loader a bit to
let the catalog finish all initialisation.

Note that the instantiation of legacy Hive UDFs doesn't seem
necessary in the catalog, we can get all relevant info from
the class. Generic UDFs do need to be instantiated to be able
to call initialize().

Testing:
- added new classes to load in test UDFs and loaded these
  in constructor / initialize()
- ran the Hive UDF ee tests

Merge conflicts in branch-4.1:
- HiveJavaFunctionFactoryImpl.java ignores the case for GenericUDF
- Ignores changes in GenericImportsNearbyClassesUdf.java

Change-Id: If16e38b8fc3b2577a5d32104ea9e6948b9562e24
Reviewed-on: http://gerrit.cloudera.org:8080/18611
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-on: http://gerrit.cloudera.org:8080/19009
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>