IMPALA-11358: Fixed Kudu table's missing comment
If Kudu-HMS integration is enabled, Kudu creates the table in HMS too,
which was missing the comment field. Added the code to forward the
comment field to Kudu during creation.

Testing:

Added a test to verify the comment is present when the intergration is
enabled.
Reenabled several kudu tests as IMPALA-8751 (and follow ups) fixed the
hive3 notification incompatibility.

Change-Id: Idf66f8b4679b00da6693a27fed79b04e8f6afb55
Reviewed-on: http://gerrit.cloudera.org:8080/18627
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>