IMPALA-11366: Fix impala-shell build failure if the build date contains Non-ASCII characters
The version string is generated in bin/save-version.sh where the
BUILD_TIME is the result of `date` command. It could contain Non-ASCII
characters depending on the locale. The date string is then used in
generating impala_build_version.py. So we need to explicitly set the
script's encoding to utf-8.

Change-Id: I596121e08a52a4bc6d8668cf7e8b61b6c34eb4b9
Reviewed-on: http://gerrit.cloudera.org:8080/18632
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Gergely Fürnstáhl <gfurnstahl@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>